# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OpenAPI Documentation using Typescript and Express using SwaggerJsDoc",
    "version": "1.0.0",
    "description": "This is a OpenAPI application made with Express. It retrieves data from local db.json file and displays it in a Swagger UI",
    "license": {
      "name": "Licensed Under MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    }
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Retrieve a list of users from db.json.",
        "description": "Retrieve a list of users from db.json. Can be used to populate a list of fake users when prototyping or testing an API.",
        "responses": {
          "200": {
            "description": "A list of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "The user ID.",
                            "example": "417d"
                          },
                          "firstname": {
                            "type": "string",
                            "description": "The user's name.",
                            "example": "Thirtha"
                          },
                          "lastname": {
                            "type": "string",
                            "description": "The user's last name.",
                            "example": "Roy"
                          },
                          "username": {
                            "type": "string",
                            "description": "The user's username.",
                            "example": "thirtha.roy"
                          },
                          "password": {
                            "type": "string",
                            "description": "The user's password.",
                            "example": "tirtha@123"
                          },
                          "email": {
                            "type": "string",
                            "description": "The user's email.",
                            "example": "tirtha@hyland.com"
                          },
                          "phone": {
                            "type": "integer",
                            "description": "The user's phone number.",
                            "example": 1234567890
                          },
                          "dateOfBirth": {
                            "type": "string",
                            "description": "The user's date of birth.",
                            "example": "1998-08-25T00:00:00.000Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a user.",
        "description": "Create a user. Can be used to create a new user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstname": {
                    "type": "string",
                    "description": "The user's name.",
                    "example": "string"
                  },
                  "lastname": {
                    "type": "string",
                    "description": "The user's last name.",
                    "example": "string"
                  },
                  "username": {
                    "type": "string",
                    "description": "The user's username.",
                    "example": "string"
                  },
                  "password": {
                    "type": "string",
                    "description": "The user's password.",
                    "example": "string"
                  },
                  "email": {
                    "type": "string",
                    "description": "The user's email.",
                    "example": "string"
                  },
                  "phone": {
                    "type": "integer",
                    "description": "The user's phone number.",
                    "example": 0
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "description": "The user's date of birth.",
                    "example": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "The user ID.",
                          "example": "517p"
                        },
                        "firstname": {
                          "type": "string",
                          "description": "The user's name.",
                          "example": "Emily"
                        },
                        "lastname": {
                          "type": "string",
                          "description": "The user's last name.",
                          "example": "Slack"
                        },
                        "username": {
                          "type": "string",
                          "description": "The user's username.",
                          "example": "emily.slack"
                        },
                        "password": {
                          "type": "string",
                          "description": "The user's password.",
                          "example": "emily@123"
                        },
                        "email": {
                          "type": "string",
                          "description": "The user's email.",
                          "example": "emily@hyland.com"
                        },
                        "phone": {
                          "type": "integer",
                          "description": "The user's phone number.",
                          "example": 9874561230
                        },
                        "dateOfBirth": {
                          "type": "string",
                          "description": "The user's date of birth.",
                          "example": "1990-08-25T00:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Retrieve a single user.",
        "description": "Retrieve a single user. Can be used to populate a user profile when prototyping or testing an API.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID."
          }
        ],
        "responses": {
          "200": {
            "description": "A single user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The user ID.",
                      "example": "417d"
                    },
                    "firstname": {
                      "type": "string",
                      "description": "The user's name.",
                      "example": "Thirtha"
                    },
                    "lastname": {
                      "type": "string",
                      "description": "The user's last name.",
                      "example": "Roy"
                    },
                    "username": {
                      "type": "string",
                      "description": "The user's username.",
                      "example": "thirtha.roy"
                    },
                    "password": {
                      "type": "string",
                      "description": "The user's password.",
                      "example": "tirtha@123"
                    },
                    "email": {
                      "type": "string",
                      "description": "The user's email.",
                      "example": "tirtha@hyland.com"
                    },
                    "phone": {
                      "type": "integer",
                      "description": "The user's phone number.",
                      "example": 1234567890
                    },
                    "dateOfBirth": {
                      "type": "string",
                      "description": "The user's date of birth.",
                      "example": "1998-08-25T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}